Add -Werror when testing -m compiler flags.
As with -f flags, some compilers need -Werror to force the compiler to
fail when passed an unknown -m flag.

Also, if it's not a -W or -f or -m flag, don't forget to add it to
CFLAGS and/or CXXFLAGS.

And propagate a comment to the C++ flag testing.

Change-Id: I53e0e59b34b10b9477c60ddd4f2e047b3b8be77f
Reviewed-on: https://code.wireshark.org/review/2361
Reviewed-by: Guy Harris <guy@alum.mit.edu>