We don't need -Qunused-arguments with Clang
check_c_compiler_flag now checks for the message Clang prints for -f and
-m flags it doesn't handle ("argument unused during compilation:
'-{flag}'"), so the checks for it now should fail properly during
testing, causing us not to use the flag in question.  This means we
don't need to suppress that warning, as we shouldn't be getting it.

Change-Id: Ieb9657f9e2cee2f357acd52725199d78d2dad80f
Reviewed-on: https://code.wireshark.org/review/2401
Reviewed-by: Guy Harris <guy@alum.mit.edu>