Allow expert info to show only items in the display filter.
This was started by Chris Maynard before Qt was announced and I just polished it off for inclusion in GTK version of Wireshark.

This also can be used as input into the Qt version of the Expert Info "dialog" since it hasn't been written yet.  Personally I like the "template" the Qt statistics dialog has with the display filter built it.  I think that would work well for the Expert Info dialog as well.

bug:1860
Change-Id: Icaada6e7900f22b0a3d97c2a5656edfd8d8c8b7f
Reviewed-on: https://code.wireshark.org/review/3035
Reviewed-by: Christopher Maynard <Christopher.Maynard@gtech.com>
Reviewed-by: Michael Mann <mmann78@netscape.net>