Treat CR-LF at the end of a line like LF.
That way, if we're reading a Windows-format file on UN*X, we handle it
the same way we'd handle a UN*X-format file.

This handles bug 10272 for the cfilter and dfilter file; there are other
configuration files that may need code changes as well.

While we're at it, don't hand non-ASCII characters to isspace().

Change-Id: Ie4ecd2c9601545a60b4d3326c6782cfff093b156
Reviewed-on: https://code.wireshark.org/review/3046
Reviewed-by: Guy Harris <guy@alum.mit.edu>