Our users are not particpants in a comedy sketch.
In way too many places in the GTK+ code we wait for the user to do
something wrong, leap out from behind the bushes, yell "AHA!", and hit
them in the face with a herring^W error dialog.

We really need to not do that. We also shouldn't carry that behavior
over to the Qt UI.

If the user shouldn't press a button it should be disabled. Do so for
the "Compile BPF" and "Start" buttons.

Change-Id: Iad9c15951d5ae640ea82ca74d5f0c26e88bd2554
Reviewed-on: https://code.wireshark.org/review/3156
Reviewed-by: Gerald Combs <gerald@wireshark.org>