Use windeployqt with NSIS.
If we find windeployqt, use it to create a DLL manifest for NSIS.
Fall back to the previous manual lists if we don't. This assumes that
your copy of windeployqt works.

New bug: The Windows installer is huge when using the official Digia DLLs
(around 43 MB for the win 64 package).

Change-Id: I1f662381b7a2c5474595b05177dbdfd3138c7e60
Reviewed-on: https://code.wireshark.org/review/3253
Reviewed-by: Gerald Combs <gerald@wireshark.org>