Hack to fix RTP player progress line position
PortAudio seems to return 0 as value for Pa_GetStreamTime() on
a PulseAudio system using the ALSA host API. This is known for 3 years,
but still has not been fixed. As this function basically gives us a
monotonically increasing clock in seconds as double, just use the Glib
library function for that. I don't know whether this adds much latency,
but it is certainly better than breaking the progress line (and also
the related auto-scroll).

While at it, split the idx calculation and explain what each part does
and add modelines. Remove invalid _U_ annotations.

 [1]: http://music.columbia.edu/pipermail/portaudio/2011-June/012452.html

Bug: 10307
Change-Id: I67fff531e0577add4b9b76052e286fb3fd5d01ff
Reviewed-on: https://code.wireshark.org/review/3260
Petri-Dish: Evan Huus <eapache@gmail.com>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Evan Huus <eapache@gmail.com>