Fix uninitialized session ticket
If the tvb contained too few data,
ssl_dissect_hnd_hello_ext_session_ticket would then allocate
session_ticket, but not initialize the contents. Fix this by adding a
check for the TVB length.

The same is done for ssl_dissect_hnd_new_ses_ticket. That might, or
might not, be necessary as proto_tree_add_item() is called with the
range. When tree is NULL, ssl is usually NULL too. For clarity (and to
avoid surprises in the future), add it anyway.

Bug: 10330
Change-Id: I469e97542542aaef4cbd660086bedf92ba1c0b6e
Reviewed-on: https://code.wireshark.org/review/3309
Reviewed-by: Evan Huus <eapache@gmail.com>
Reviewed-by: Anders Broman <a.broman58@gmail.com>