Remove the old wireshark binary during RPM upgrades.
If we have a pre-alternatives wireshark binary out there (e.g., because
we're upgrading from wireshark-1.10 to wireshark-1.12), get rid of it
in the %pre script.  Otherwise the binary gets left there (because it's
listed as a %ghost in the new spec file).

Change-Id: I5828fcff486af3d269bad8dfe544c7659179bf5e
Reviewed-on: https://code.wireshark.org/review/3332
Reviewed-by: Jeff Morriss <jeff.morriss.ws@gmail.com>
(cherry picked from commit c93297c750b99ac487f1f787896959ca43f64c72)
Reviewed-on: https://code.wireshark.org/review/3440