CompiledFilterOutput dialog fixes and updates.
Apply a layout to the "Compiled Filter Output" dialog. Render the filter
as monospaced text.

Don't add an "expert OK" icon to the list. It doesn't seem to indicate
any sort of status.

Use a "Close" button instead of "OK". According to

http://uxmovement.com/buttons/why-the-ok-button-is-no-longer-okay/

"OK" implies we're going to perform some sort of action.

Add a "Copy" button while we're here.

Change-Id: Ia63f3032916e322fa9f4f2d465140b128b4f8078
Reviewed-on: https://code.wireshark.org/review/3338
Reviewed-by: Gerald Combs <gerald@wireshark.org>