Rename buffer_ routines to tagbuffer_ to avoid name collisions.
In particular, collisions with the buffer_ routines in libwsutil; in the
trunk, we renamed *those* routines, but that breaks API/ABI
compatibility, so, in the 1.12 branch, we rename these routines, which
adds a difference between this code and the Lrexlib release that we'd
rather not have if we can avoid it.

This should fix bug 10332.

Change-Id: Ib37a50713acb880fb79adacc91f7bb49959be97e
Reviewed-on: https://code.wireshark.org/review/3352
Reviewed-by: Guy Harris <guy@alum.mit.edu>