Make Python mandatory to build on Windows
While nominally optional, building without it doesn't work, so just make it
mandatory.

The verify_tools target already passes $(PYTHON) to win-setup.sh --appverify;
we just need to supply a default value for PYTHON, so that $(PYTHON) doesn't
expand to an empty string. It's also convenient for the developer if Python
can be automatically found on the PATH.

Change-Id: I4f54695625b74c5b4c758ef1e3bc9ed4467db514
Reviewed-on: https://code.wireshark.org/review/3391
Petri-Dish: Jörg Mayer <jmayer@loplof.de>
Reviewed-by: Jörg Mayer <jmayer@loplof.de>
Tested-by: Jörg Mayer <jmayer@loplof.de>
Reviewed-by: Anders Broman <a.broman58@gmail.com>