When doing the distcheck cmake check only verify that cmake runs, don't verify that it can build Wireshark.
The purpose of the cmake check is to ensure we are distributing all the of
necessary cmake files.  We don't want to get into the business of giving cmake
all the options required to actually build Wireshark (e.g., which version of
Qt is installed and where).

This also means that distcheck doesn't compile Wireshark twice.

Change-Id: I62f5c42df7c710f7695bd020b0dda80e54435829
Reviewed-on: https://code.wireshark.org/review/3432
Reviewed-by: Evan Huus <eapache@gmail.com>