Clean up the CRC-10 code.
Have the wsutil routine just accumulate the stuff from the buffer handed
to us.  Have the IUUP dissector deal with the extra stuff.  Add a
update_crc10_by_bytes_tvb() routine, which is passed a tvbuff, offset,
and length, and use that rather than using tvb_get_ptr() in dissectors.

Change-Id: Iadd0823c764080e60d1339abb94d2e19150eabfe
Reviewed-on: https://code.wireshark.org/review/3509
Reviewed-by: Guy Harris <guy@alum.mit.edu>