An NLPID of 0xc0 refers to TRILL.
This will, at least, make it show up as the meaning of a field
containing an NLPID; whether it's actually used to say "a TRILL packet
follows" is another matter.

At least partially addresses bug 10382, and fully addresses it if it's
never actually used to say "a TRILL packet follows".

Change-Id: Ic9dc8ec432532bde692fad3aa9cc97a3109aa634
Reviewed-on: https://code.wireshark.org/review/3602
Reviewed-by: Guy Harris <guy@alum.mit.edu>