Fix Ceph packet length determination.
The authentication string was not being factored in on the length reply.
Also there was an issue with different banners.  Now the banner length
must match what we expect or the packet is rejected.  If the banner
length changes the protocol is different and we won't be able to parse
it anyways.

Change-Id: I0c1a7379edaa203042486a0e6f9ce3642427da99
Reviewed-on: https://code.wireshark.org/review/3710
Reviewed-by: Michael Mann <mmann78@netscape.net>