9P2000.L also has the UID field in tattach.
See, for example:

	http://code.google.com/p/diod/wiki/protocol#attach,_auth_--_messages_to_establish_a_connection

and

	http://code.google.com/p/diod/wiki/protocol#Overview

where the latter says

	9P2000.L consists of a subset of the canonical 9P2000
	operations, the 9P2000.u attach and auth messages, plus new
	operations designed to map to the Linux VFS in a straightforward
	way.

so tattach and tauth are the same as 9P2000.u.

Don't throw a dissector assertion of there's data at the end of the
packet; in the particular case that found this bug, yes, it was a
dissector bug, but it could also be a malformed packet, and we should
not throw dissector assertions unless we know *for certain* that the
problem is a dissector bug rather than a malformed packet.

Change-Id: I5ef5d837bccb4e7c4844d31a138aa55a85ec1b60
Reviewed-on: https://code.wireshark.org/review/3758
Reviewed-by: Guy Harris <guy@alum.mit.edu>