Exclude MA USB TransferAck's from USB conversation
the usb_trans_info struct only understands Requests & Responses.
TransferAcks are currently treated as responses, which throws off
the Request/Response pairing & dissection.

This patch fixes tracking in the case when a TransferAck & TransferReq
are in the same TCP packet. It does not fix the case where there are
multiple TransferReq's or TransferResp's in 1 TCP packet.

Eventually I want to revamp usb_trans_info structure to support:
 -multiple requests/responses per transaction (IE: retries)
 -TransferAck tracking
 -handling of multiple requests/responses in 1 TCP packet

Change-Id: I09be6fecf55a40b4ea2a4f27bd753f3991ccabbd
Reviewed-on: https://code.wireshark.org/review/3866
Reviewed-by: Pascal Quantin <pascal.quantin@gmail.com>